app/socket_server/sockets.py
import socketio

# create a Socket.IO server
sio = socketio.AsyncServer(
    async_mode='asgi',
    cors_allowed_origins=["http://localhost:3001"]
)

@sio.event
async def connect(sid, environ, auth):
    print('connect ', sid)

@sio.event
async def disconnect(sid):
    print('disconnect ', sid)

# wrap with ASGI application
sio_app = socketio.ASGIApp(
    socketio_server=sio
)