requirements.txt
fastapi[all]
uvicorn[standard]
python-socketio